Update error handling in authentications controller. Rename files
