+++ app/views/beyond_canvas/authentications/new.html.erb
