+++ beyond_canvas.gemspec
@@ -26,7 +26,7 @@ Gem::Specification.new do |spec|
 
   spec.required_ruby_version = '>= 2.5'
 
-  spec.add_dependency 'beyond_api',      '~> 0.8'
+  spec.add_dependency 'beyond_api',      '~> 0.11'
   spec.add_dependency 'bourbon',         '~> 5.1'
   spec.add_dependency 'colorize',        '~> 0.8'
   spec.add_dependency 'http-accept',     '~> 2.1'
