+++ lib/generators/beyond_canvas/install/templates/en.yml
@@ -1,0 +1,4 @@
+en:
+  beyond_canvas:
+    authentications:
+      failure: 'Shop could not be saved'
